+++ src/lib/db/db/querys.ts
@@ -1,5 +1,5 @@
 import { getDb } from ".";
-import { eq, count, sql, and } from "drizzle-orm";
+import { eq, count, sql, and, inArray } from "drizzle-orm";
 import { playlists, playlistsSongs } from "./schema";
 
 export async function getPlaylists() {
@@ -63,4 +63,35 @@ export async function removeSongFromPlaylist(
 export async function removeSongFromAllPlaylists(songId: string){
   const db = await getDb();
   await db.delete(playlistsSongs).where(eq(playlistsSongs.songId, songId))
+}
+
+
+export async function removeManySongsFromPlaylist(playlistId: number, songIds: string[]) {
+  if (songIds.length === 0) return;
+  const db = await getDb();
+  await db.delete(playlistsSongs).where(
+    and(
+      eq(playlistsSongs.playlistId, playlistId),
+      inArray(playlistsSongs.songId, songIds),
+    ),
+  );
+}
+
+export async function removeManySongFromAllPlaylists(songIds: string[]){
+  if (songIds.length === 0) return;
+  const db = await getDb();
+  await db.delete(playlistsSongs).where(inArray(playlistsSongs.songId, songIds))
+}
+
+export async function addManySongsToPlaylists(playlistId: number, songsId: string[]){
+if (songsId.length === 0) return;
+
+  const db = await getDb();
+
+  const valuesToInsert = songsId.map((songId) => ({
+    playlistId: playlistId,
+    songId: songId,
+  }));
+
+  await db.insert(playlistsSongs).values(valuesToInsert);
 }