+++ CHANGELOG.md
@@ -6,8 +6,10 @@ El formato está basado en [Keep a Changelog](https://keepachangelog.com/es-ES/1
 
 ---
 ## [Unreleased]
+
 ### Added
 - Eliminacion de canciones del almacenamiento local
+- Seleccion multiple de canciones.
 # [0.2.4] - 2026-08-05
 - Dialog para mostrar errores al usuario en tiempo de ejecucion